Kill-process andd add-exit-hook
Introduced a function for killing processes,
and a function for adding an exit hook.

The reason for this functionality is that a Data Explorer process
started by Femlisp always hung around even after Femlisp was
terminated - eating up 100% CPU time.  Now, this process is
explicitly killed when Femlisp terminates with a normal exit.